Refactor: #21 의 리뷰요청 사항 반영(Repository 통합)
