+++ data/expense/src/main/java/com/kappzzang/jeongsan/di/ExpenseRepositoryModule.kt
@@ -1,13 +1,11 @@
 package com.kappzzang.jeongsan.di
 
 import com.kappzzang.jeongsan.repository.ExpenseDetailRepository
-import com.kappzzang.jeongsan.repository.ExpenseListRepository
 import com.kappzzang.jeongsan.repository.ExpenseRepository
 import com.kappzzang.jeongsan.repository.ReceiptRepository
 import com.kappzzang.jeongsan.repository.TransferRepository
 import com.kappzzang.jeongsan.repositoryimpl.ExpenseDetailRepositoryImpl
-import com.kappzzang.jeongsan.repositoryimpl.ExpenseListFakeRepositoryImpl
-import com.kappzzang.jeongsan.repositoryimpl.ExpenseRepositoryImpl
+import com.kappzzang.jeongsan.repositoryimpl.ExpenseFakeRepositoryImpl
 import com.kappzzang.jeongsan.repositoryimpl.ReceiptRepositoryImpl
 import com.kappzzang.jeongsan.repositoryimpl.TransferRepositoryImpl
 import dagger.Binds
@@ -21,21 +19,15 @@ import javax.inject.Singleton
 abstract class ExpenseRepositoryModule {
     @Binds
     @Singleton
-    abstract fun bindExpensePageRepository(
-        expenseListRepository: ExpenseListFakeRepositoryImpl
-    ): ExpenseListRepository
+    abstract fun bindExpenseRepository(
+        expenseListRepository: ExpenseFakeRepositoryImpl
+    ): ExpenseRepository
 
     @Binds
     @Singleton
     abstract fun bindExpenseDetailRepository(
         expenseDetailRepositoryImpl: ExpenseDetailRepositoryImpl
     ): ExpenseDetailRepository
-
-    @Binds
-    @Singleton
-    abstract fun bindExpenseRepository(
-        expenseRepositoryImpl: ExpenseRepositoryImpl
-    ): ExpenseRepository
 
     @Binds
     @Singleton
