+++ domain/group/src/main/java/com/kappzzang/jeongsan/usecase/GetInviteInfoUseCase.kt
@@ -13,7 +13,7 @@ class GetInviteInfoUseCase(private val memberRepository: MemberRepository) {
                 MemberItem(
                     id = "id$i",
                     name = "멤버 이름$i",
-                    profileImageURL = "",
+                    profileImageUrl = "",
                     isInvited = i % 2 != 0
                 )
             )
